Attempted reset flicker fix and code notes
Tried to fix the part where it flickers on button hold by moving the
held check above clicked. Didn’t work. Will have to go back to example
code for the encoder library and figure out what HELD means in this
context.

Made notes about a couple of blocks that need to be cleaned up with
TODO in the source.

1) The presets are being loaded in a switch/case block by the same for
loop each time. Need to pull the repeated code into a function call and
pass by reference of the presets variable.

2) The button read code is used too many places and should be moved to
the top of the loop and just compared after that.